Change proposal for project specification
Added terrain generation to will do.
Moved grass animation to will do.
Moved shadow mapping to might do.

We have some experience with terrain generation from the labs, generating this is almost a prerequisite to grass. (unless we want it on a flat ground of a loaded model which can also work).
I don't think grass animation will be difficult if we use the technique mentioned.
I think shadow maps would be nice but more of a parallel track.